+++ src/theory/strings/theory_strings_preprocess.h
@@ -91,14 +91,6 @@ class StringsPreprocess {
    * visited stores a cache of previous results.
    */
   Node simplifyRec(Node t, std::vector<Node>& asserts);
-  /**
-   * Make internal quantified formula with bound variable list bvl and body.
-   * Internally, we get a node corresponding to marking a quantified formula as
-   * an "internal" one. This node is provided as the third argument of the
-   * FORALL returned by this method. This ensures that E-matching is not applied
-   * to the quantified formula.
-   */
-  static Node mkForallInternal(Node bvl, Node body);
 };
 
 }  // namespace strings
